+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - feat(services): add log level selection sub menu to each service menu and sync it with persistent storage - closes #164
 - test(services): set limits for the number of registered listeners and event handlers after all services are loaded in `test_all_services_register`
 - refactor(core): wait for all service threads to join before running cleanup functions for dbus connections and the logger
+- feat(services): apply log levels set in service settings for each service using `logging.Filter`
 
 ## Version 1.3.0